Fix write size

+++ clients/libfs-write/src/main.c
@@ -16,15 +16,15 @@ int main(int argc, char** argv) {
     }
 
     // TODO: verify content lenght works correctly
-    int result = libfs_write(atoi(argv[1]), (u8*)argv[2], (u32)strlen(argv[2]) + 1);
+    ssize_t result = libfs_write(atoi(argv[1]), (u8*)argv[2], (u32)strlen(argv[2]) + 1);
 
     if (result < 0) {
         LOG_LIBFS_ERRNO("libfs_write failed");
         return 1;
     }
 
-    LOG_INFO("libfs_write succeeded: %d", result);
-    printf("%d", result);
+    LOG_INFO("libfs_write succeeded: %ld", result);
+    printf("%ld", result);
 
     return 0;
 }
